Create BaseManifestReader from ManifestReader.
