+++ core/src/test/java/org/apache/iceberg/TestManifestReader.java
@@ -44,10 +44,9 @@ public class TestManifestReader extends TableTestBase {
   }
 
   @Test
-  @SuppressWarnings("deprecation")
   public void testManifestReaderWithEmptyInheritableMetadata() throws IOException {
-    ManifestFile manifest = writeManifest("manifest.avro", manifestEntry(Status.EXISTING, 1000L, FILE_A));
-    try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
+    ManifestFile manifest = writeManifest(1000L, manifestEntry(Status.EXISTING, 1000L, FILE_A));
+    try (ManifestReader reader = ManifestFiles.read(manifest, FILE_IO)) {
       ManifestEntry entry = Iterables.getOnlyElement(reader.entries());
       Assert.assertEquals(Status.EXISTING, entry.status());
       Assert.assertEquals(FILE_A.path(), entry.file().path());
@@ -56,22 +55,18 @@ public class TestManifestReader extends TableTestBase {
   }
 
   @Test
-  @SuppressWarnings("deprecation")
   public void testInvalidUsage() throws IOException {
     ManifestFile manifest = writeManifest(FILE_A, FILE_B);
-    try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
-      AssertHelpers.assertThrows(
-          "Should not be possible to read manifest without explicit snapshot ids and inheritable metadata",
-          IllegalArgumentException.class, "must have explicit snapshot ids",
-          () -> Iterables.getOnlyElement(reader.entries()));
-    }
+    AssertHelpers.assertThrows(
+        "Should not be possible to read manifest without explicit snapshot ids and inheritable metadata",
+        IllegalArgumentException.class, "Cannot read from ManifestFile with null (unassigned) snapshot ID",
+        () -> ManifestFiles.read(manifest, FILE_IO));
   }
 
   @Test
-  @SuppressWarnings("deprecation")
   public void testManifestReaderWithPartitionMetadata() throws IOException {
-    ManifestFile manifest = writeManifest("manifest.avro", manifestEntry(Status.EXISTING, 123L, FILE_A));
-    try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
+    ManifestFile manifest = writeManifest(1000L, manifestEntry(Status.EXISTING, 123L, FILE_A));
+    try (ManifestReader reader = ManifestFiles.read(manifest, FILE_IO)) {
       ManifestEntry entry = Iterables.getOnlyElement(reader.entries());
       Assert.assertEquals(123L, (long) entry.snapshotId());
 
@@ -84,7 +79,6 @@ public class TestManifestReader extends TableTestBase {
   }
 
   @Test
-  @SuppressWarnings("deprecation")
   public void testManifestReaderWithUpdatedPartitionMetadataForV1Table() throws IOException {
     PartitionSpec spec = PartitionSpec.builderFor(table.schema())
         .bucket("id", 8)
@@ -92,8 +86,8 @@ public class TestManifestReader extends TableTestBase {
         .build();
     table.ops().commit(table.ops().current(), table.ops().current().updatePartitionSpec(spec));
 
-    ManifestFile manifest = writeManifest("manifest.avro", manifestEntry(Status.EXISTING, 123L, FILE_A));
-    try (ManifestReader reader = ManifestReader.read(FILE_IO.newInputFile(manifest.path()))) {
+    ManifestFile manifest = writeManifest(1000L, manifestEntry(Status.EXISTING, 123L, FILE_A));
+    try (ManifestReader reader = ManifestFiles.read(manifest, FILE_IO)) {
       ManifestEntry entry = Iterables.getOnlyElement(reader.entries());
       Assert.assertEquals(123L, (long) entry.snapshotId());
